+++ Modules/Intent.Modules.Common.CSharp/Builder/CSharpField.cs
@@ -1,6 +1,9 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
+using Intent.Metadata.Models;
 using Intent.Modules.Common.CSharp.Builder.InterfaceWrappers;
+using Intent.Modules.Common.CSharp.VisualStudio;
 
 #nullable enable
 
@@ -39,7 +42,8 @@ public class CSharpField : CSharpMember<CSharpField>, ICSharpField
         AccessModifier = "private ";
         Type = type;
         Name = name;
-        Parent = @class;    
+        Parent = @class;
+        File = @class?.File; // can be null because of CSharpInterfaceProperty :(
     }
 
     internal static CSharpField CreateFieldOmittedFromRender(string type, string name, ICSharpCodeContext @class, string? value)
@@ -121,6 +125,29 @@ public class CSharpField : CSharpMember<CSharpField>, ICSharpField
     {
         Assignment = value;
         return this;
+    }
+
+    public CSharpField WithInstantiation(ITypeReference model)
+    {
+        if (File?.Template.OutputTarget.GetProject().GetLanguageVersion().Major >= 12 && model is not null && model.IsCollection)
+        {
+            Assignment = "[]";
+            return this;
+        }
+
+        if (File?.Template.OutputTarget.GetProject().GetLanguageVersion().Major < 12 && model is not null && model.IsCollection)
+        {
+            var propertyType = CSharpTypeParser.Parse(Type);
+            var concreteImplementation = propertyType.GetCollectionImplementationType();
+
+            Assignment = propertyType is CSharpTypeGeneric ? $"new {File.Template.GetTypeName(model, File.Template.UseType($"{concreteImplementation}<{{0}}>")).Replace("?", "")}()" :
+                $"new {File.Template.GetTypeName(model, File.Template.UseType($"{concreteImplementation}")).Replace("?", "")}()";
+
+            return this;
+        }
+
+        return this;
+
     }
 
     public override string GetText(string indentation)
@@ -240,6 +267,8 @@ public class CSharpField : CSharpMember<CSharpField>, ICSharpField
     {
         return _wrapping.ProtectedReadOnly();
     }
+
+    ICSharpField ICSharpField.WithInstantiation(ITypeReference model) => _wrapping.WithInstantiation(model);
 
     #endregion
 }