+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -5,6 +5,7 @@
 - Fixed: Return types of IAsyncEnumerable are now not automatically wrapped in Task/ValueTask when method is marked as async
 - Fixed: Constructor Mappings not using `Select(...).ToList()` pattern for collections.
 - Improvement: Added support for mapping of inherited DTO fields
+- Improvement: Added WithInstantiation method to CSharpProperty and CSharpField, with LangVersion aware collection support
 
 ### Version 3.8.1